requirements.txt
flask
PyMuPDF
faiss-cpu
requests
tqdm
sentence-transformers
torch 